requirements.txt
aiogram==3.13.0
aiosqlite==0.20.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
openpyxl==3.1.5
asyncpg
alembic
apscheduler
matplotlib
